+++ Nikesaurus/assets/Shaders/batched_texture.vert
@@ -16,7 +16,7 @@ layout (location = 6) in vec3 a_model_to_ndc_2;
 layout (location=8) in vec2 framesize;
 layout (location=9) in vec2 uv_offset;
 
-layout (location=11) in int a_blend;		// bool
+layout (location=11) in float a_blend;		// bool
 layout (location=12) in vec4 a_color;
 layout (location=13) in float a_blend_intensity;
 
@@ -25,7 +25,7 @@ layout (location=0) out vec2 f_texcoord;
 layout (location=1) flat out float f_sampler_idx;
 //layout (location=2) out vec2 dbg_framesize;
 
-layout (location=3) flat out int f_blend;
+layout (location=3) flat out float f_blend;
 layout (location=4) flat out vec4 f_color;
 layout (location=5) flat out float f_blend_intensity;
